jobapp/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
authentication==1.1.0
beautifulsoup4==4.11.1
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudinary==1.29.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.3
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-debug-toolbar==3.3.0
django-js-asset==2.0.0
django-oauth-toolkit==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
facebook-sdk==3.1.0
Flask==2.1.2
google==3.0.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.1
httplib2==0.20.4
idna==3.3
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
passlib==1.7.4
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.4.0
PyMySQL==1.1.0
pyparsing==3.0.8
pytz==2022.1
PyYAML==6.0.1
redis==4.3.1
reportlab==4.0.4
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.7.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.16.2
